requirements.txt
adjustText==0.7.3
anyio==3.2.1
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.2.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bce-python-sdk==0.8.61
bleach==3.2.1
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.4
cfgv==3.3.0
chardet==3.0.4
click==8.0.1
cmocean==2.0
colorama==0.4.4
colormath==3.0.0
colorspacious==1.1.2
contextvars==2.4
cryptography==3.1.1
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.2
Django==2.2.5
django-cleanup==4.0.0
django-extensions==2.2.6
django-nested-admin==3.2.4
entrypoints==0.3
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-Babel==2.0.0
future==0.18.2
gast==0.5.0
gunicorn==20.0.4
identify==2.2.11
idna==2.8
imageio==2.9.0
imgaug==0.4.0
immutables==0.15
importlib-metadata==2.0.0
importlib-resources==5.2.0
iniconfig==1.1.1
ipykernel==5.3.4
ipynb==0.5.1
ipyparallel==6.3.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.15.1
Jinja2==3.0.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
kiwisolver==1.3.0
llvmlite==0.34.0
lmdb==1.2.1
MarkupSafe==2.0.1
matlabengineforpython==R2019b
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.5.0
nbclassic==0.3.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
networkx==2.5
nodeenv==1.6.0
notebook==6.1.4
numba==0.51.2
numpy==1.19.2
olefile==0.46
opencv-contrib-python==4.5.3.56
packaging==20.4
paddleocr==2.0.6
paddlepaddle==2.1.1
pandas==1.1.3
pandocfilters==1.4.2
parso==0.8.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pip==20.2.4
platformdirs==2.0.2
pluggy==0.13.1
pre-commit==2.13.0
progressbar2==3.37.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.17.3
ptyprocess==0.6.0
py==1.9.0
pyclipper==1.3.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pyflakes==2.3.1
Pygments==2.7.2
pyldpc==0.7.9
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
pytesseract==0.3.8
pytest==0.0.0
pytest-runner==5.2
python-dateutil==2.8.1
python-Levenshtein==0.12.2
python-monkey-business==1.0.0
python-utils==2.4.0
pytz==2020.1
pyusb==1.0.2
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==19.0.2
qrcode==6.1
requests==2.22.0
requests-unixsocket==0.2.0
scikit-image==0.17.2
scikit-learn==0.22
scipy==1.3.2
seaborn==0.9.0
segno==1.3.3
Send2Trash==1.5.0
setuptools==50.3.0.post20201006
Shapely==1.7.1
shellcheck-py==0.7.2.1
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
statsmodels==0.10.1
terminado==0.9.1
testpath==0.4.4
tifffile==2020.9.3
toml==0.10.1
tornado==6.1
tqdm==4.61.2
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.25.11
virtualenv==20.6.0
visualdl==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
wheel==0.35.1
zipp==3.4.0
